requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118

# torch==2.2.0+cu118
# torchvision==0.17.0+cu118
# torchaudio==2.2.0+cu118

numpy==1.26.3
matplotlib==3.8.4
pandas==2.2.2
scipy==1.13.0
pytorch_lightning==1.5.10
tqdm>=4.41.0
pyyaml>=5.1
future>=0.17.1
tensorboard>=2.2.0
